# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Boot.dev Guided Project",
        "title": "RSS-Aggregator",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/feed_follows": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all FeedFollows for authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed_follows"
                ],
                "summary": "Gets all FeedFollows for authenticated user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/FeedFollow"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new FeedFollow for authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed_follows"
                ],
                "summary": "Creates a FeedFollow for authenticated user",
                "parameters": [
                    {
                        "description": "feed id",
                        "name": "feedfollow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/FeedFollowCreateParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/FeedFollow"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/feed_follows/{feedFollowID}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes specific FeedFollow for authenticated user",
                "tags": [
                    "feed_follows"
                ],
                "summary": "Deletes specific FeedFollow for authenticated user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "feed uuid",
                        "name": "feedFollowID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/feeds": {
            "get": {
                "description": "Get all feeds",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeds"
                ],
                "summary": "Gets all feeds",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Feed"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new feed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeds"
                ],
                "summary": "Creates a feed",
                "parameters": [
                    {
                        "description": "feed name and url",
                        "name": "feed",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/FeedCreateParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/FeedAndFeedFollows"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Gets current user information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Gets current user information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Creates a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Creates a user",
                "parameters": [
                    {
                        "description": "user name",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserCreateParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "Feed": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-10-05T00:49:11.897425Z"
                },
                "id": {
                    "type": "string",
                    "example": "4bda8766-37a2-46e7-ad95-1c4038b3bcfe"
                },
                "name": {
                    "type": "string",
                    "example": "The Example Blog"
                },
                "update_at": {
                    "type": "string",
                    "example": "2023-10-05T00:49:11.897425Z"
                },
                "url": {
                    "type": "string",
                    "example": "https://example.blog"
                },
                "user_id": {
                    "type": "string",
                    "example": "03f474bc-2651-4a2f-ab3e-9525d1ee40bf"
                }
            }
        },
        "FeedAndFeedFollows": {
            "type": "object",
            "properties": {
                "feed": {
                    "$ref": "#/definitions/Feed"
                },
                "feed_follow": {
                    "$ref": "#/definitions/FeedFollow"
                }
            }
        },
        "FeedCreateParams": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "The Example Blog"
                },
                "url": {
                    "type": "string",
                    "example": "https://example.blog"
                }
            }
        },
        "FeedFollow": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-10-05T03:57:38.673971Z"
                },
                "feed_id": {
                    "type": "string",
                    "example": "4bda8766-37a2-46e7-ad95-1c4038b3bcfe"
                },
                "id": {
                    "type": "string",
                    "example": "1323e287-5e9d-4999-8373-35293f38c9d8"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-10-05T03:57:38.673971Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "03f474bc-2651-4a2f-ab3e-9525d1ee40bf"
                }
            }
        },
        "FeedFollowCreateParams": {
            "type": "object",
            "properties": {
                "feed_id": {
                    "type": "string",
                    "example": "26296df9-59ce-4fbe-8c7d-f3397fcb91ce"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string",
                    "example": "333a7712b7edb73922f9e5c0421fe224443247aa80da93b0409377281b38207c"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-10-05T00:47:49.007148Z"
                },
                "id": {
                    "type": "string",
                    "example": "03f474bc-2651-4a2f-ab3e-9525d1ee40bf"
                },
                "name": {
                    "type": "string",
                    "example": "John"
                },
                "update_at": {
                    "type": "string",
                    "example": "2023-10-05T00:47:49.007148Z"
                }
            }
        },
        "UserCreateParams": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "John"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}